move setting files into post-schedule folder
